+++ src/main/java/org/example/tropicaliasapi/service/UserService.java
@@ -41,7 +41,7 @@ public class UserService {
     }
 
 
-    public User getByID(int id) {
+    public User getByID(Long id) {
         return userRepository.findById(id).orElse(null);
     }
